Hotfix

+++ pages/logout.js
@@ -3,13 +3,20 @@ import { useRouter } from 'next/router'
 import { useEffect, useState } from 'react'
 import MainLayout from '../components/MainLayout'
 import useSWR from 'swr'
+import { checkToken } from '../controllers/auth/public_auth'
+import { useSearchParams } from 'next/navigation'
+import withAuth from '../components/ProtectedRoute'
 
-function Logout () {
+
+function Logout (props) {
     const router = useRouter()
 
     const [isClient, setIsClient] = useState(false)
 
+    const pageParams = useSearchParams()
+
     async function logout(url) {
+
         return await fetch(url, {
                     method: 'POST',
                     headers: {
@@ -26,33 +33,34 @@ function Logout () {
 }
 
     const {data} = useSWR(`${process.env.NEXT_PUBLIC_API_URL}/rest-auth/logout/`, logout)
+
+   
     
     useEffect(() => {
 
         setIsClient(true)
 
-        // const cookieCutter = require('cookie-cutter')
-        // cookieCutter.set('access_token', '', "{}", { expires: new Date(0), httpOnly: false })
+        
+                if (!data?.error && !data?.detail /*&& props?.response*/) {
+                    if (typeof window !== 'undefined') {
+                        window.sessionStorage.removeItem('user')
+                        window.localStorage.removeItem('user')
+                        window.localStorage.clear()
+                        window.document.cookie = 'access_token=; expires=Thu, 01 Jan 1970 00:00:00 GMT;'
+                        router.push('/')
 
-        // setTimeout(() => {
-            if (!data?.error && !data?.detail) {
-                if (typeof window !== 'undefined') {
-                    window.sessionStorage.removeItem('user')
-                    window.localStorage.removeItem('user')
-                    window.localStorage.clear()
-                    window.document.cookie = 'access_token=; expires=Thu, 01 Jan 1970 00:00:00 GMT;'
-                    router.push('/')
-
+                    } else {
+                        const cookieCutter = require('cookie-cutter')
+                        cookieCutter.set('access_token', '', "{}", { expires: new Date(0), httpOnly: false })
+                    }
                 } else {
-                    const cookieCutter = require('cookie-cutter')
-                    cookieCutter.set('access_token', '', "{}", { expires: new Date(0), httpOnly: false })
+                        // window.localStorage.clear()  
+                        router.replace(pageParams.get('previous_path'))
+
+
                 }
-            } else {
-                    window.localStorage.clear()
-                    router.replace('/')
-
-
-            }
+            
+    
         // }, 1000);
 
     }, [])
@@ -80,23 +88,41 @@ function Logout () {
     }
 }
 
-// Logout.getInitialProps = async () => {
-//     // const cookieCutter = require('cookie-cutter')
-//     const API_URL = process.env.NEXT_PUBLIC_API_URL
-//     return fetch(API_URL + '/rest-auth/logout/', {
-//         method: 'POST',
-//         headers: {
-//             'Content-Type': 'application/json'
+
+// export async function getServerSideProps(ctx) {
+
+//     const token = (await checkToken(ctx?.req, ctx?.res))?.token
+
+//     const payload = {
+//         token: token,
+//         client_id: process.env.CLIENT_ID,
+//         client_secret: process.env.CLIENT_SECRET
+//     }
+
+//     let response = null
+
+//     if(token) {
+//         response = await (await fetch(`${process.env.TOKEN_URL.replace('token','revoke_token')}`, {
+//             headers: {
+//                 'Content-Type': 'application/json',
+//                 'Authorization': `Bearer ${token}`
+//             },
+//             method: 'POST',
+//             body: JSON.stringify(payload)
+//         })).json()
+//     }
+
+
+//     console.log({response})
+    
+
+
+//     return {
+//         props: {
+//             response
 //         }
-//     })
-//         .then((response) => {
-           
-//             if (response.error || response.detail) {
-//                 return { error: true, ...response }
-//             }
-//             return response
-//         })
+//     }
 // }
 
-export default Logout
+export default withAuth(Logout)
